pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.64.0",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "cryptography>=45.0.6",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "marshmallow>=4.0.0",
    "marshmallow-sqlalchemy>=1.4.2",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pymysql>=1.1.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "sendgrid>=6.12.4",
    "stripe>=12.4.0",
]
